requirements.txt
keras==2.4.3
tensorflow-cpu==2.5.0
Flask==0.11.1
Jinja2==2.10
MarkupSafe==0.23
Werkzeug==0.11.11
click==6.7
gunicorn==19.6.0
itsdangerous==0.24
requests
numpy
clint
oauth2client
scipy
flask_restful
flask-jsonpify
pandas
rq
timeout-decorator
xlrd
botnoi==0.2.1
matplotlib
gensim
scikit-learn
sklearn